backend/requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
web3>=5.24.0
python-dotenv>=0.19.0
pydantic>=2.0.0
pydantic-settings>=2.0.0  # 添加这行
sqlalchemy>=1.4.23
alembic==1.7.1
pytest==6.2.5
httpx==0.19.0
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.5
aiohttp==3.8.1